pull out DDatabase from DatomicData source file
put DDatabase into its own source file
